Make GET(Detail), PUT
JSON 파싱에서 로직을 자체적으로 만들어보려다 너무 무거워져서 스트림 및 람다식 사용하는 건 다른 분의 코드 내용을 참조하였습니다.

PUT/PATCH를 실행하면 어느 부분에선가 인코딩/디코딩 문제가 발생하여 수정 중에 있습니다.
